layers merging into one file
