+++ official/vision/beta/projects/yolo/modeling/layers/dark_tiny_test.py
@@ -3,7 +3,7 @@ import tensorflow.keras as ks
 import numpy as np
 from absl.testing import parameterized
 
-from official.vision.beta.projects.yolo.modeling import building_blocks as nn_blocks
+from official.vision.beta.projects.yolo.modeling import layers as nn_blocks
 
 
 class DarkTinyTest(tf.test.TestCase, parameterized.TestCase):
